criando servico de amenidades

+++ book-room/quarto/src/main/java/br/com/book/room/quarto/infrastructure/database/postgres/repository/LocalidadeRepository.java
@@ -3,6 +3,6 @@ package br.com.book.room.quarto.infrastructure.database.postgres.repository;
 import br.com.book.room.quarto.infrastructure.database.postgres.entity.LocalidadeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalidadeRepository extends JpaRepository<LocalidadeEntity, Integer> {
+public interface LocalidadeRepository extends JpaRepository<LocalidadeEntity, Long> {
 
 }
